requirements.txt
datasets==3.3.2
matplotlib==3.10.0
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.5.1
torchvision==0.20.1
